Make useAsyncOperation lint friendly
Currently, it's not possible to use the exhaustive-deps lint with useAsyncOperation while also using the ability to return undefined to signal RemoteData.NOT_ASKED.
This is due to the lint expecting an inline function as the first argument in all cases.
This change allows the caller to use a synchronous function to work around that.